Updated dependencies
- Updated PHP minimum supported version to 5.6.
- Updated all module and component dependencies to latest stable releases; if v3
  releases are available, they are supported via `||` constraints.
- Moved `Module` class into `src/` tree, where it belongs, and:
  - Removed unneeded `getAutoloaderConfig()` implementation.
  - Updated to use short array syntax.
  - Updated to use `::class` notation.
  - Ensured view strategy is attached to event manager in a forwards compatible
    way.
