+++ .gitignore
@@ -1,2 +1,1 @@
 vendor/
-composer.lock
